+++ crates/assistant_slash_commands/Cargo.toml
@@ -35,7 +35,6 @@ rope.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 smol.workspace = true
-terminal_view.workspace = true
 text.workspace = true
 toml.workspace = true
 ui.workspace = true
