+++ crates/eval/Cargo.toml
@@ -61,6 +61,7 @@ settings.workspace = true
 shellexpand.workspace = true
 smol.workspace = true
 telemetry.workspace = true
+terminal_view.workspace = true
 toml.workspace = true
 unindent.workspace = true
 util.workspace = true
